Moves CollectionStrategyTest and CascadeTest to Builder\Field namespace
